requirements.txt
numpy
faiss-cpu
ollama
PyPDF2
streamlit